Example from NN_Viz.py
In this image, the black dots represent the particles, and the arrows represent whether they are spin-up or spin-down. The red bonds are satisfied (low energy) when they connect two parallel spins, and the blue bonds are satisfied when they connect anti-parallel spins. The bonds with white lines in the middle represent bonds that the algorithm has identified will be unsatisfied in a ground energy state.